refactor: split governance instructions from lib.rs into governance.rs
- Extract governance-related functionality into separate governance.rs module
- Keep all pyth operations in lib.rs as requested
- Maintain idiomatic Rust module organization
- All code compiles successfully with only minor warnings
